backend/requirements.txt
# Le Sésame Backend - Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
greenlet>=3.0.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# LLM Integration
mistralai>=1.0.0
openai>=1.12.0
httpx>=0.26.0

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-mistralai>=0.2.0
langchain-google-genai>=2.0.0
langchain-aws>=0.2.0

# Redis for sessions/caching
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0

# Async support
aiofiles>=23.2.1
